A basic websocket server in Go
It only receives messages and doesn't forward them to the customer for
now.

We're consuming the WebSocket message readers in a custom way (no
ioutil.ReadAll() :)) so that keeping the amount of RAM per connection
is possible. Not that the implementation is rather dummy for now, we
could improve it by growing the data slice by chunks of n bytes until
we reach EOF or the maximum allowed size for a given message.