Add a docker-compose based dev. env.
To make it easy and fast to hack on the project, on any docker-enabled
workstation.

Also, note that the golang binary must be statically compiled since it
is run in a "FROM scratch" docker image. The (long) command to do so
will live in a Makefile, we won't have to type it all the time.