Add a Makefile to the project
It pulls dependencies & builds the Golang backend in a build container.
Therefore, there's no need to worry about running the wrong version of
Golang locally.

The Makefile can also be used to rebuild the devenv in one command when
(and only if) necessary as well as commands to tail its logs and
destroy it.